Improve tooltip and add custom word functionality
Make tooltip appear on click and maintain visibility.

Make tooltip disappear on click if it is visibile.

Make all visible tooltips disappear when clicking
to appear a hidden tooltip.

Make all visible tooltips disappear when clicking
anywhere on the document that is not an action button
for a tooltip.

Create button and text box for retrieving custom words.

Create clipboard copy button for copying retrieved
custom words.